Add expand_keys_diff property and behavior.
The expand_keys behavior is cool, but it currently deletes the entire
tree represented by a file and recreates it for any change made in a
file. This is not desireable for any large number of keys for two
reasons:

1. When only a subset of the tree changes, we result in a large number
of writes.
2. Watches will fire when keys that have not really changes are
modified.

This commit adds a setting called "expand_keys_diff" that can be set at
the repo level. When set to true, then a diff is performed between the
resulting tree of the file and the existing tree in consul. Only the
necessary add/update/delete operations are performed, and thus both
concerns above are eliminated.

The question of consistency is a red herring as in general practice the
replication lag between consul instances it negligible, and good
practice would dictate that no other client be modifying the keys
mirrored by git2consul.