Apply changes based on code review.
Comments have been updated to better articulate the reason for doing
either a diff or a delete of the current tree.

The 'toString' operation is no longer applied in the "render_obj"
function, but rather moved to the "diff_kvs" function where it is
actually a concern. I.e. the types and values for the "value" attribute
of the KV objects must match.

Further, the "diff_kvs" function was refactored to have better
performance for large trees. By removing keys from the "existing_kvs"
array as we find them (i.e. when we know we won't need to delete them)
we reduce the number of operations we need to construct the
"delete_kvs" array. In fact, the "delete_kvs" array is simply a copy of
any elements still remaining in "existing_kvs" because we logically
know at this point that they were not in the "candidate_kvs" array, and
thus must be deleted from the current tree.