Make test suite actually exit with proper exit code.
Current test suite always exits with zero because it's piped through
bunyan.  Since bunyan is always successful and it's the last command
on the pipe chain we get it's exit code.

By moving the command into a shell script and setting the pipefail
option for bash we will exit with the proper exit code.  This means
that Travis CI will correctly determine test failures.