Add support for certificates for secure mode.
This allows for an optional Certifying Authority certificate as well as
a Client certificate and key for use with secure mode.  Allowing
git2consul to be used with servers that have verify_incoming turned on.

Consul client library had to be upgraded to pick up a new enough version
of the node consul client library and it's dependency papi to support
this.