Use --no-renames with doing git diff.
git 2.x seems to automatically have renames on and git2consul
doesn't know how to handle the R status.  Might be better
to handle all the statuses, but this appears to make it work with
newer versions of git without further changes so going with this.

Tested this with 1.9.1 and 2.14.1.

Noticed this since Travis CI now has 2.13.0